src/backend/requirements.txt
sentry-sdk[fastapi]
fastapi
fastapi[standard]
python-dotenv
llama-index-graph-stores-neo4j
alive-progress
RateMyProfessorAPI
llama-index-llms-gemini
llama-index-llms-azure-openai
llama-index-llms-vertex